Start using scoped enums in many places for readability.
Instead of foo use UHDM_OBJECT_TYPE::foo if something is of
that type. This does not cover all places yet, but it will allow
us to make UHDM_OBJECT_TYPE an enum class at some point and improve
type-safety.

+++ templates/Serializer_save.cpp
@@ -59,7 +59,7 @@ struct Serializer::SaveAdapter {
     if (obj->VpiParent() != nullptr) {
       ::ObjIndexType::Builder vpiParentBuilder = builder.getVpiParent();
       vpiParentBuilder.setIndex(GetId(obj->VpiParent(), idMap));
-      vpiParentBuilder.setType(obj->VpiParent()->UhdmType());
+      vpiParentBuilder.setType(static_cast<uint32_t>(obj->VpiParent()->UhdmType()));
     }
     builder.setVpiFile((RawSymbolId)obj->GetSerializer()->symbolMaker.Make(obj->VpiFile()));
     builder.setVpiLineNo(obj->VpiLineNo());
